# queez-flask-api/static/swagger.json
{   
    "openapi": "3.0.0",
    "info": {
            "description": "Landing Quarry API",
            "version": "1.0.0",
            "title": "Landing Quarry Service"
    },    
    "servers": [
        {
            
            "url": "http://127.0.0.1:5000"
        },
        {
            
            "url": "http://127.0.0.1:5000/api/auth/quary"
        },
        {
            "url": "http://127.0.0.1:5000/api/auth/command"
        },
        {
            "url": "http://127.0.0.1:5000/api/landing/quary"
        },
        {
            "url": "http://127.0.0.1:5000/api/landing/command"
        },
        {
            "url": "http://127.0.0.1:5000/api/advertisement/quary"
        },
        {
            "url": "http://127.0.0.1:5000/api/advertisement/command"
        }
    ],
    "paths": {
...
        "/api/advertisement/quary/get_advertisement_by_id/{advertisement_id}": {
            "get": {
                "tags": ["Advertisement"],
                "summary": "Get Advertisement by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "advertisement_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"  
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/advertisement/quary/get_advertisement_by_filter": {
            "get": {
                "tags": ["Advertisement"],
                "summary": "Get Advertisement by Filter",
                "parameters": [
                    {
                        "in": "query",
                        "name": "filter_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "customer_id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object" 
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/customerportal/quary/get_pinned_items": {
            "get": {
                "tags": ["CustomerPortal"],
                "summary": "Get Pinned Items",
                "parameters": [
                    {
                        "in": "query",
                        "name": "customer_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object" 
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customerportal/quary/get_saved_filters": {
            "get": {
                "tags": ["CustomerPortal"],
                "summary": "Get Saved Filters",
                "parameters": [
                    {
                        "in": "query",
                        "name": "customer_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object" 
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}